Beef up the Music class and implement events.
Improves the Music class to generate clicks, beats and bars at regular intervals
(depending on when it is configured to do so). The actual configuration is
currently done by BaseModel.FirstUpdate, but this is actually the responsibility
of the AudioEngine component and should be moved there when ready.

The current demo/toy consists of a single active turret, rotating 18 degrees
every beat, in a 4/4 60bpm piece (so each beat comes one second after the last).

Note that we are now using TimeSpan rather than GameTime to represent the game's
current time: this class is more appropriate for our purposes, and while
GameTime is not an immutable snapshot, TimeSpan is.