Implement upgrades, improve ShotPatterns.
The UpgradeReady event was unused, so I've changed its use to something more
useful. It is now used to signal that an upgrade menu has been opened - it is
implied that the player is ready to upgrade if this happens. UpgradeDone
indicates that the menu should be closed: the player has finished upgrading or
has cancelled the menu.